feat: create a separate user provider component
