+++ src/hocs/RequireAuth.tsx
@@ -2,7 +2,7 @@ import React, { useContext, type FC } from 'react';
 import { Navigate } from 'react-router-dom';
 import { type ChildrenProps } from './types';
 import { pagePathnames } from 'router/pagePathnames';
-import { UserContext } from 'App';
+import { UserContext } from 'store/userContext';
 
 const RequireAuth: FC<ChildrenProps> = ({ children }) => {
   const { isUserLoggedIn } = useContext(UserContext);
